# src/docs/swagger.yaml
definitions:
  github_com_alielmi98_golang-todo-list-api_api_dto.CreateToDoRequest:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
  github_com_alielmi98_golang-todo-list-api_api_dto.LoginByUsernameRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
  github_com_alielmi98_golang-todo-list-api_api_dto.PagedList-github_com_alielmi98_golang-todo-list-api_api_dto_ToDoResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPrevPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.ToDoResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  github_com_alielmi98_golang-todo-list-api_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties: true
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        additionalProperties:
          type: string
        type: object
    type: object
  github_com_alielmi98_golang-todo-list-api_api_dto.RegisterUserByUsernameRequest:
    properties:
      email:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
  github_com_alielmi98_golang-todo-list-api_api_dto.ToDoResponse:
    properties:
      completed:
        type: boolean
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
  github_com_alielmi98_golang-todo-list-api_api_dto.UpdateToDoRequest:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        type: string
    required:
    - description
    - title
    type: object
  github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.ResultCode'
      success:
        type: boolean
    type: object
  github_com_alielmi98_golang-todo-list-api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    - 50003
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
    - InvalidInputError
info:
  contact: {}
paths:
  /v1/todo/:
    post:
      consumes:
      - application/json
      description: Create a new todo job
      parameters:
      - description: Create a todo
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.CreateToDoRequest'
      responses:
        "201":
          description: ToDo response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.ToDoResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a todo
      tags:
      - Todo
  /v1/todo/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo job by Id
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "201":
          description: Todo response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  type: string
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a todo
      tags:
      - Todo
    get:
      consumes:
      - application/json
      description: Get a todo job by Id
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "201":
          description: Todo response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.ToDoResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a todo
      tags:
      - Todo
    put:
      consumes:
      - application/json
      description: Update a todo job
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a todo
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.UpdateToDoRequest'
      responses:
        "201":
          description: Todo response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.ToDoResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a todo
      tags:
      - Todo
  /v1/todo/filter:
    post:
      consumes:
      - application/json
      description: Retrieve a list of ToDos based on filter criteria with pagination
        support
      parameters:
      - description: Pagination and filter input
        in: body
        name: paginationInput
        required: true
        schema:
          $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.PaginationInputWithFilter'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with list of ToDos
          schema:
            allOf:
            - $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.PagedList-github_com_alielmi98_golang-todo-list-api_api_dto_ToDoResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "404":
          description: No ToDos found
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get ToDos by filter with pagination
      tags:
      - Todo
  /v1/tokens/refresh-token:
    post:
      consumes:
      - application/json
      description: RefreshToken
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      summary: RefreshToken
      tags:
      - Users
  /v1/users/login-by-username:
    post:
      consumes:
      - application/json
      description: LoginByUsername
      parameters:
      - description: LoginByUsernameRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.LoginByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      summary: LoginByUsername
      tags:
      - Users
  /v1/users/register-by-username:
    post:
      consumes:
      - application/json
      description: RegisterByUsername
      parameters:
      - description: RegisterUserByUsernameRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_dto.RegisterUserByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_alielmi98_golang-todo-list-api_api_helper.BaseHttpResponse'
      summary: RegisterByUsername
      tags:
      - Users
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"